test(admin): split monolith admin unit spec to match page migration
